Split tests into multiple tests Move ref tests from test to ref directory
