+++ test/ref/gather.cpp
@@ -56,157 +56,161 @@ TEST_CASE(gather_non_std_test)
     }
 }
 
-TEST_CASE(gather_test)
+TEST_CASE(gather_test_1)
 {
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3 * 3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
-        std::vector<int> indices{0, 2};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = 0;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data(4 * 5);
-        std::vector<float> golden = {0.5f, 1.5f, 2.5f, 6.5f, 7.5f, 8.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3 * 3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
+    std::vector<int> indices{0, 2};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = 0;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data(4 * 5);
+    std::vector<float> golden = {0.5f, 1.5f, 2.5f, 6.5f, 7.5f, 8.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+TEST_CASE(gather_test_2)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3 * 3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
-        std::vector<int> indices{-3, -1};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = 0;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data(4 * 5);
-        std::vector<float> golden = {0.5f, 1.5f, 2.5f, 6.5f, 7.5f, 8.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3 * 3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
+    std::vector<int> indices{-3, -1};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = 0;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data(4 * 5);
+    std::vector<float> golden = {0.5f, 1.5f, 2.5f, 6.5f, 7.5f, 8.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+TEST_CASE(gather_test_3)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3 * 3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
-        std::vector<int> indices{0, 2};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = 1;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data(4 * 5);
-        std::vector<float> golden = {0.5f, 2.5f, 3.5f, 5.5f, 6.5f, 8.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3 * 3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
+    std::vector<int> indices{0, 2};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = 1;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data(4 * 5);
+    std::vector<float> golden = {0.5f, 2.5f, 3.5f, 5.5f, 6.5f, 8.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+TEST_CASE(gather_test_4)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3 * 3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
-        std::vector<int> indices{0, 2};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = -1;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data(4 * 5);
-        std::vector<float> golden = {0.5f, 2.5f, 3.5f, 5.5f, 6.5f, 8.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3 * 3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    migraphx::shape s_indices{migraphx::shape::int32_type, {1, 2}};
+    std::vector<int> indices{0, 2};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = -1;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data(4 * 5);
+    std::vector<float> golden = {0.5f, 2.5f, 3.5f, 5.5f, 6.5f, 8.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+TEST_CASE(gather_test_5)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3 * 3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        // scalar index
-        migraphx::shape s_indices{migraphx::shape::int32_type};
-        std::vector<int> indices{0};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = -1;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data{};
-        std::vector<float> golden = {0.5f, 3.5f, 6.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3 * 3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    // scalar index
+    migraphx::shape s_indices{migraphx::shape::int32_type};
+    std::vector<int> indices{0};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = -1;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data{};
+    std::vector<float> golden = {0.5f, 3.5f, 6.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+TEST_CASE(gather_test_6)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3 * 3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        // scalar index
-        migraphx::shape s_indices{migraphx::shape::int32_type};
-        std::vector<int> indices{-3};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = -1;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data{};
-        std::vector<float> golden = {0.5f, 3.5f, 6.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3 * 3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    // scalar index
+    migraphx::shape s_indices{migraphx::shape::int32_type};
+    std::vector<int> indices{-3};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = -1;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data{};
+    std::vector<float> golden = {0.5f, 3.5f, 6.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
+TEST_CASE(gather_test_7)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
 
-        std::vector<float> data(3);
-        std::iota(data.begin(), data.end(), 0.5);
-        migraphx::shape s{migraphx::shape::float_type, {3}};
-        auto a0 = mm->add_literal(migraphx::literal{s, data});
-        // scalar index
-        migraphx::shape s_indices{migraphx::shape::int32_type};
-        std::vector<int> indices{0};
-        auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
-        int axis = -1;
-        mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> res_data{};
-        std::vector<float> golden = {0.5f};
-        result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
-        EXPECT(migraphx::verify::verify_range(res_data, golden));
-    }
+    std::vector<float> data(3);
+    std::iota(data.begin(), data.end(), 0.5);
+    migraphx::shape s{migraphx::shape::float_type, {3}};
+    auto a0 = mm->add_literal(migraphx::literal{s, data});
+    // scalar index
+    migraphx::shape s_indices{migraphx::shape::int32_type};
+    std::vector<int> indices{0};
+    auto a1  = mm->add_literal(migraphx::literal{s_indices, indices});
+    int axis = -1;
+    mm->add_instruction(migraphx::make_op("gather", {{"axis", axis}}), a0, a1);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> res_data{};
+    std::vector<float> golden = {0.5f};
+    result.visit([&](auto output) { res_data.assign(output.begin(), output.end()); });
+    EXPECT(migraphx::verify::verify_range(res_data, golden));
 }
 
 TEST_CASE(gather_dyn_test0)
