+++ test/ref/prefix_scan_sum.cpp
@@ -70,193 +70,191 @@ TEST_CASE(prefix_scan_sum_dyn_1d)
     EXPECT(results_vector == gold);
 }
 
-TEST_CASE(prefix_scan_sum_2d)
+TEST_CASE(prefix_scan_sum_2d_1)
 {
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 0}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0, 2.0, 3.0, 2.0, 4.0, 6.0, 3.0, 6.0, 9.0};
-        EXPECT(results_vector == gold);
-    }
-
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 1}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0, 3.0, 6.0, 1.0, 3.0, 6.0, 1.0, 3.0, 6.0};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 0}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0, 2.0, 3.0, 2.0, 4.0, 6.0, 3.0, 6.0, 9.0};
+    EXPECT(results_vector == gold);
 }
 
-TEST_CASE(prefix_scan_sum_3d)
+TEST_CASE(prefix_scan_sum_2d_2)
 {
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 0}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0,
-                                2.0,
-                                3.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                2.0,
-                                4.0,
-                                6.0};
-        EXPECT(results_vector == gold);
-    }
-
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 1}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                3.0,
-                                6.0,
-                                9.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                3.0,
-                                6.0,
-                                9.0};
-        EXPECT(results_vector == gold);
-    }
-
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 2}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 1}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0, 3.0, 6.0, 1.0, 3.0, 6.0, 1.0, 3.0, 6.0};
+    EXPECT(results_vector == gold);
 }
 
-TEST_CASE(prefix_scan_sum_exclusive)
+TEST_CASE(prefix_scan_sum_3d_1)
 {
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {8}};
-        auto input = migraphx::literal{s, {1, 2, 3, 4, 1, 2, 3, 4}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 0}, {"exclusive", true}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{0.0, 1.0, 3.0, 6.0, 10.0, 11.0, 13.0, 16.0};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 0}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0,
+                            2.0,
+                            3.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            2.0,
+                            4.0,
+                            6.0};
+    EXPECT(results_vector == gold);
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", 1}, {"exclusive", true}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{0.0,
-                                0.0,
-                                0.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                0.0,
-                                0.0,
-                                0.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0};
-        EXPECT(results_vector == gold);
-    }
+TEST_CASE(prefix_scan_sum_3d_2)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 1}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            3.0,
+                            6.0,
+                            9.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            3.0,
+                            6.0,
+                            9.0};
+    EXPECT(results_vector == gold);
+}
+
+TEST_CASE(prefix_scan_sum_3d_3)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 2}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0};
+    EXPECT(results_vector == gold);
+}
+
+TEST_CASE(prefix_scan_sum_exclusive_1)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {8}};
+    auto input = migraphx::literal{s, {1, 2, 3, 4, 1, 2, 3, 4}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 0}, {"exclusive", true}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{0.0, 1.0, 3.0, 6.0, 10.0, 11.0, 13.0, 16.0};
+    EXPECT(results_vector == gold);
+}
+
+TEST_CASE(prefix_scan_sum_exclusive_2)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", 1}, {"exclusive", true}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{0.0,
+                            0.0,
+                            0.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            0.0,
+                            0.0,
+                            0.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0};
+    EXPECT(results_vector == gold);
 }
 
 TEST_CASE(prefix_scan_sum_exclusive_reverse)
@@ -277,143 +275,142 @@ TEST_CASE(prefix_scan_sum_exclusive_reverse)
     EXPECT(results_vector == gold);
 }
 
-TEST_CASE(prefix_scan_sum_negative_axis)
+TEST_CASE(prefix_scan_sum_negative_axis_1)
 {
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", -3}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0,
-                                2.0,
-                                3.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                2.0,
-                                4.0,
-                                6.0};
-        EXPECT(results_vector == gold);
-    }
-
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", -2}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                3.0,
-                                6.0,
-                                9.0,
-                                1.0,
-                                2.0,
-                                3.0,
-                                2.0,
-                                4.0,
-                                6.0,
-                                3.0,
-                                6.0,
-                                9.0};
-        EXPECT(results_vector == gold);
-    }
-
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
-        auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum", {{"axis", -1}, {"exclusive", false}}), l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0,
-                                1.0,
-                                3.0,
-                                6.0};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", -3}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0,
+                            2.0,
+                            3.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            2.0,
+                            4.0,
+                            6.0};
+    EXPECT(results_vector == gold);
 }
 
-TEST_CASE(prefix_scan_sum_reverse)
+TEST_CASE(prefix_scan_sum_negative_axis_2)
 {
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {8}};
-        auto input = migraphx::literal{s, {1, 2, 3, 4, 1, 2, 3, 4}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum",
-                              {{"axis", 0}, {"exclusive", false}, {"reverse", true}}),
-            l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{20.0, 19.0, 17.0, 14.0, 10.0, 9.0, 7.0, 4.0};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", -2}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            3.0,
+                            6.0,
+                            9.0,
+                            1.0,
+                            2.0,
+                            3.0,
+                            2.0,
+                            4.0,
+                            6.0,
+                            3.0,
+                            6.0,
+                            9.0};
+    EXPECT(results_vector == gold);
+}
 
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 2, 2}};
-        auto input = migraphx::literal{s, {1, 2, 3, 4, 1, 2, 3, 4}};
-        auto l0    = mm->add_literal(input);
-        mm->add_instruction(
-            migraphx::make_op("prefix_scan_sum",
-                              {{"axis", 0}, {"exclusive", false}, {"reverse", true}}),
-            l0);
-        p.compile(migraphx::make_target("ref"));
-        auto result = p.eval({}).back();
-        std::vector<float> results_vector;
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{2.0, 4.0, 6.0, 8.0, 1.0, 2.0, 3.0, 4.0};
-        EXPECT(results_vector == gold);
-    }
+TEST_CASE(prefix_scan_sum_negative_axis_3)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 3, 3}};
+    auto input = migraphx::literal{s, {1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum", {{"axis", -1}, {"exclusive", false}}), l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0,
+                            1.0,
+                            3.0,
+                            6.0};
+    EXPECT(results_vector == gold);
+}
+
+TEST_CASE(prefix_scan_sum_reverse_1)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {8}};
+    auto input = migraphx::literal{s, {1, 2, 3, 4, 1, 2, 3, 4}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum",
+                          {{"axis", 0}, {"exclusive", false}, {"reverse", true}}),
+        l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{20.0, 19.0, 17.0, 14.0, 10.0, 9.0, 7.0, 4.0};
+    EXPECT(results_vector == gold);
+}
+
+TEST_CASE(prefix_scan_sum_reverse_2)
+{
+    migraphx::program p;
+    auto* mm = p.get_main_module();
+    migraphx::shape s{migraphx::shape::float_type, {2, 2, 2}};
+    auto input = migraphx::literal{s, {1, 2, 3, 4, 1, 2, 3, 4}};
+    auto l0    = mm->add_literal(input);
+    mm->add_instruction(
+        migraphx::make_op("prefix_scan_sum",
+                          {{"axis", 0}, {"exclusive", false}, {"reverse", true}}),
+        l0);
+    p.compile(migraphx::make_target("ref"));
+    auto result = p.eval({}).back();
+    std::vector<float> results_vector;
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{2.0, 4.0, 6.0, 8.0, 1.0, 2.0, 3.0, 4.0};
+    EXPECT(results_vector == gold);
 }
