+++ test/ref/quantizelinear.cpp
@@ -30,57 +30,56 @@
 
 #include <test.hpp>
 
-TEST_CASE(quantizelinear)
+TEST_CASE(quantizelinear_1)
 {
-    {
-        migraphx::shape xs{migraphx::shape::float_type, {2, 3, 3}};
-        std::vector<float> xv = {
-            -300, 600, 129, -1000, 4, 3, -6, 600, 550, -300, 600, 129, -1000, 4, 3, -6, 600, 550};
-        migraphx::shape ss{migraphx::shape::float_type, {2, 3, 3}};
-        std::vector<float> sv = {2, 2, 2, 4, 4, 4, 6, 6, 6, 2, 2, 2, 4, 4, 4, 6, 6, 6};
-        migraphx::shape zs{migraphx::shape::int8_type, {2, 3, 3}};
-        std::vector<uint8_t> zv = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
-        auto create_program     = [&]() {
-            migraphx::program p;
-            auto* mm = p.get_main_module();
-            auto x   = mm->add_literal(xs, xv);
-            auto s   = mm->add_literal(ss, sv);
-            auto z   = mm->add_literal(zs, zv);
-            mm->add_instruction(migraphx::make_op("quantizelinear"), x, s, z);
-            return p;
-        };
+    migraphx::shape xs{migraphx::shape::float_type, {2, 3, 3}};
+    std::vector<float> xv = {
+        -300, 600, 129, -1000, 4, 3, -6, 600, 550, -300, 600, 129, -1000, 4, 3, -6, 600, 550};
+    migraphx::shape ss{migraphx::shape::float_type, {2, 3, 3}};
+    std::vector<float> sv = {2, 2, 2, 4, 4, 4, 6, 6, 6, 2, 2, 2, 4, 4, 4, 6, 6, 6};
+    migraphx::shape zs{migraphx::shape::int8_type, {2, 3, 3}};
+    std::vector<uint8_t> zv = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+    auto create_program     = [&]() {
+        migraphx::program p;
+        auto* mm = p.get_main_module();
+        auto x   = mm->add_literal(xs, xv);
+        auto s   = mm->add_literal(ss, sv);
+        auto z   = mm->add_literal(zs, zv);
+        mm->add_instruction(migraphx::make_op("quantizelinear"), x, s, z);
+        return p;
+    };
 
-        migraphx::program p1 = create_program();
-        p1.compile(migraphx::make_target("ref"));
-        auto result = p1.eval({}).back();
-        std::vector<float> results_vector(18);
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{
-            -128, 127, 65, -128, 1, 1, -1, 100, 92, -128, 127, 65, -128, 1, 1, -1, 100, 92};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p1 = create_program();
+    p1.compile(migraphx::make_target("ref"));
+    auto result = p1.eval({}).back();
+    std::vector<float> results_vector(18);
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{
+        -128, 127, 65, -128, 1, 1, -1, 100, 92, -128, 127, 65, -128, 1, 1, -1, 100, 92};
+    EXPECT(results_vector == gold);
+}
 
-    {
-        migraphx::shape xs{migraphx::shape::float_type, {2, 3, 3}};
-        std::vector<float> xv = {
-            -300, 600, 129, -1000, 4, 3, -6, 600, 550, -300, 600, 129, -1000, 4, 3, -6, 600, 550};
-        migraphx::shape ss{migraphx::shape::float_type, {2, 3, 3}};
-        std::vector<float> sv = {2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2};
-        auto create_program   = [&]() {
-            migraphx::program p;
-            auto* mm = p.get_main_module();
-            auto x   = mm->add_literal(xs, xv);
-            auto s   = mm->add_literal(ss, sv);
-            mm->add_instruction(migraphx::make_op("quantizelinear"), x, s);
-            return p;
-        };
+TEST_CASE(quantizelinear_2)
+{
+    migraphx::shape xs{migraphx::shape::float_type, {2, 3, 3}};
+    std::vector<float> xv = {
+        -300, 600, 129, -1000, 4, 3, -6, 600, 550, -300, 600, 129, -1000, 4, 3, -6, 600, 550};
+    migraphx::shape ss{migraphx::shape::float_type, {2, 3, 3}};
+    std::vector<float> sv = {2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2};
+    auto create_program   = [&]() {
+        migraphx::program p;
+        auto* mm = p.get_main_module();
+        auto x   = mm->add_literal(xs, xv);
+        auto s   = mm->add_literal(ss, sv);
+        mm->add_instruction(migraphx::make_op("quantizelinear"), x, s);
+        return p;
+    };
 
-        migraphx::program p1 = create_program();
-        p1.compile(migraphx::make_target("ref"));
-        auto result = p1.eval({}).back();
-        std::vector<float> results_vector(18);
-        result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
-        std::vector<float> gold{0, 255, 65, 0, 2, 2, 0, 255, 255, 0, 255, 65, 0, 2, 2, 0, 255, 255};
-        EXPECT(results_vector == gold);
-    }
+    migraphx::program p1 = create_program();
+    p1.compile(migraphx::make_target("ref"));
+    auto result = p1.eval({}).back();
+    std::vector<float> results_vector(18);
+    result.visit([&](auto output) { results_vector.assign(output.begin(), output.end()); });
+    std::vector<float> gold{0, 255, 65, 0, 2, 2, 0, 255, 255, 0, 255, 65, 0, 2, 2, 0, 255, 255};
+    EXPECT(results_vector == gold);
 }
